Added tqdm, moved BG calcs to own file
