+++ requirements.txt
@@ -17,3 +17,4 @@ recommonmark
 myst_parser
 astroquery>=0.4.2
 asdf
+tqdm
